requirements/base.txt
Django==2.0.7
django-bootstrap3==10.0.1
dj-database-url==0.5.0
dj-static==0.0.6
django-csp==3.4
django-csp-reports==1.1
django-crispy-forms==1.7.2
django-filter==2.0.0
djangorestframework==3.8.2
django-secure==1.0.1
django-toolbelt==0.0.1
django-user-sessions==1.5.3
gunicorn==19.9.0
gviz_api==1.9.0
py-bcrypt==0.4
psycopg2==2.7.5
pytz==2018.5
requests==2.19.1
static==1.1.1
whitenoise[brotli]==4.0
urllib3[secure]==1.23
future==0.16.0
coreapi==2.3.3
Markdown==2.6.11
pygments==2.2.0
